Remove unexplained call to handleObservationStatement
A BATTERY CompoundStatement can only have two types of component
inside it according to the GP2GP specification. These are:

- A CLUSTER CompoundStatement
- An ObservationStatement

The existing code within handleBatteryCompoundStatement deals with
each of those cases already. The additional code removed in this
change appears to deal with any type of CompoundStatement. Given
that the only CompoundStatement which is valid to appear here is
a CLUSTER, and the code for handleClusterCompoundStatement already
calls handleObservationStatement, there is no point to this code.

By removing this duplicate call to handleObservationStatement we
also fix the issue where Observations are being generated with
two identical "category" values of "Laboratory".

Furthermore, this code is completely untested. It is a liability in
its current state, so the removal of it reduces the chance of it
breaking and introducing more unwelcome unintended side-effects.