NIAD-2394: Populate the "no disclosure to patient" label for Blood pressure
Add `ConfidentialityService` to `BloodPressureMapper`.
Add functionality to populate the security meta when the provided `ehrComposition` contains a "NOPAT" `confidentialityCode`.
Add functionality to populate the security meta when the provided `compoundStatement` contains a "NOPAT" `confidentialityCode`.
Add functionality to populate the security meta when the diastolic or systolic `ObservationStatement` contains a "NOPAT" `confidentialityCode`.
Add Unit Tests for the above functionality.
Add Unit Test xml files for the above tests.