NIAD-3170: Medication Request Orders and Plans not generated correctly
Add Test to ensure a new `Plan` is not generated when the original `Plan` is not `Acute` (i.e. `Repeat`).
Update Test from previous commit to ensure that it generates and new `Plan` only when the original `Plan` is `Acute`.
Add / Rename test files for the above tests.
Add Test to ensure that the earliest `Order` references the original `Plan` when a new `Plan` is generated.