NIAD-3170: Acute MedicationRequest[Order]s share the same MedicationStatement
Add functionality to ensure that a new `MedicationStatement` is generated when multiple `MedicationRequest[Order]`s share the same acute `MedicationRequest[Plan]`.
Add functionality to duplicate the existing `MedicationStatement` and copy the contained values using the existing MedicationStatement.copy() method.
Add functionality to generate and set a new `id` and `identifier` for the generated `MedicationStatement` based on the ID of the associated generated `MedicationRequest[Plan]`.
Add functionality to update the `basedOn` reference to instead reference the generated `MedicationRequest[Plan]`.
Add functionality to update `effectivePeriod` to match that of `dispenseRequest.validityPeriod` from the associated `MedicationRequest[Order]`.
Add functionality to update the value for the `LastIssueDate` extension to match the `dispenseRequest.validityPeriod.start` value from the associated `MedicationRequest[Order]`.
Add unit tests for the above functionality.