NIAD-3220: Detect Referral + Document LinkSets and not create a Problem (Condition) from them
Add functionality to ConditionMapper to ensure that a `Condition` is not mapped from a `LinkSet` when it is a "ReferralRequestToExternalDocumentLinkSet."
Add unit test for the above functionality.
Add functionality to ReferralRequestMapper to set `supportingInfo` with `DocumentReferences` to when the `ReferralRequest` is referenced from a "ReferralRequestToExternalDocumentLinkSet."
Add unit test for above functionality.
Update PWTP7_vis-output.json to reflect the changes made.